// src/app/model/addEvent.ts
export interface AddEvent{
    name?: string;
    description?: string;
    start?: any;
    end?: any;
    url1?: any;
    url2?: any;
    url3?: any;
    url4?: any;
    url5?: any;
    url6?: any;
    url7?: any;
    url8?: any;
    url9?: any;
    id?: string;
    
}

export interface DisplayEvent{
    name?: string;
    description?: string;    
    start?: any;
    end?: any;
    url1?: any;
    url2?: any;
    url3?: any;
    url4?: any;
    url5?: any;
    url6?: any;
    url7?: any;
    url8?: any;
    url9?: any;
    id?: string;
    
}